Avoid GetText deprecation warning when using newer GetText
Newer versions of gettext are pickier about the Plural-Forms header
structure than before, and will ignore the nplurals behavior unless
there is 1) a rule for which plural form to support and 2) the rule
is in the expected format.

To avoid deprecation warnings, the Plural form rules need to be corrected.
Alternatively, custom implementations of the GetText.Plural behavior can
be used, especially if custom behavior is required. See
[GetText.Plural module documentation](https://hexdocs.pm/gettext/Gettext.Plural.html)

As far as I can tell, the rules used by Timex mirror the rules
implemented by the transitive dependency gettext 0.22 now has on
[Expo](https://github.com/elixir-gettext/expo), so I don't see an
immediate need for a custom implementation of Gettext.Plural.

This PR also removes an incorrectly specified plural rule for
Indonesian.

There may be some remaining issues with Russian and possibly languages
with similar pluralization rules, like Croatian, Polish, and Servian,
as several unit tests for Russian still fail.